Refactor RPC api structure
