+++ api/ethapi/backend.go
@@ -115,62 +115,6 @@ type Backend interface {
 	SccApiBackend
 }
 
-func GetAPIs(apiBackend Backend) []rpc.API {
-	nonceLock := new(AddrLocker)
-	return []rpc.API{
-		{
-			Namespace: "eth",
-			Version:   "1.0",
-			Service:   NewPublicEthereumAPI(apiBackend),
-			Public:    true,
-		}, {
-			Namespace: "eth",
-			Version:   "1.0",
-			Service:   NewPublicBlockChainAPI(apiBackend),
-			Public:    true,
-		}, {
-			Namespace: "dag",
-			Version:   "1.0",
-			Service:   NewPublicDAGChainAPI(apiBackend),
-			Public:    true,
-		}, {
-			Namespace: "eth",
-			Version:   "1.0",
-			Service:   NewPublicTransactionPoolAPI(apiBackend, nonceLock),
-			Public:    true,
-		}, {
-			Namespace: "txpool",
-			Version:   "1.0",
-			Service:   NewPublicTxPoolAPI(apiBackend),
-			Public:    true,
-		}, {
-			Namespace: "debug",
-			Version:   "1.0",
-			Service:   NewPrivateDebugAPI(apiBackend),
-		}, {
-			Namespace: "eth",
-			Version:   "1.0",
-			Service:   NewPublicAccountAPI(apiBackend.AccountManager()),
-			Public:    true,
-		}, {
-			Namespace: "personal",
-			Version:   "1.0",
-			Service:   NewPrivateAccountAPI(apiBackend, nonceLock),
-			Public:    false,
-		}, {
-			Namespace: "abft",
-			Version:   "1.0",
-			Service:   NewPublicAbftAPI(apiBackend),
-			Public:    true,
-		}, {
-			Namespace: "sonic",
-			Version:   "1.0",
-			Service:   NewPublicSccApi(apiBackend),
-			Public:    true,
-		},
-	}
-}
-
 // GetVmConfig is a utility function resolving the VM configuration for a block
 // height based on the network rules.
 func GetVmConfig(
