+++ gossip/filters/api.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/0xsoniclabs/sonic/evmcore"
 
-	"github.com/0xsoniclabs/sonic/ethapi"
+	"github.com/0xsoniclabs/sonic/api/ethapi"
 	"github.com/Fantom-foundation/lachesis-base/inter/idx"
 	ethereum "github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
